requirements.txt
anyio==4.4.0
asgiref==3.8.1
cachetools==5.4.0
Cerberus==1.3.5
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==5.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
kolo==2.33.6
mocket==3.12.8
more-itertools==10.4.0
msgpack==1.0.8
oauthlib==3.2.2
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-magic==0.4.27
#python-magic-bin==0.4.14
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==69.5.1
sniffio==1.3.1
sqlglot==25.16.0
sqlparse==0.5.1
tabulate==0.9.0
toolz==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
ulid-py==1.1.0
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.43.0
